pyproject.toml
[tool.poetry]
name = "blogbot"
version = "0.1.0"
description = "A Python-based static site generator with FastHTML web interface"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.12"
python-fasthtml = "^0.12.19"
httpx = "^0.28.0"
pydantic = "^2.9.0"
pyyaml = "^6.0.2"
markdown-it-py = "^3.0.0"
python-frontmatter = "^1.1.0"
authlib = "^1.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.21.0"
ruff = "^0.7.0"
mypy = "^1.12.0"
pre-commit = "^4.0.0"
playwright = "^1.49.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "B", "A", "C4", "PIE", "PYI", "Q", "RSE", "RET", "SIM", "TID", "UP", "YTT"]
ignore = ["F403", "F405"]  # Allow star imports for FastHTML

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]